+++ doc/changelog.rst
@@ -23,6 +23,8 @@ Contributors
 
 Added
 -----
+- EBSD.detector property storing an EBSDDetector.
+  (`#262 <https://github.com/pyxem/kikuchipy/pull/262>`_)
 - Link to Binder in README and in the notebooks for running them in the browser.
   (`#257 <https://github.com/pyxem/kikuchipy/pull/257>`_)
 - A data module with a small Nickel EBSD data set and master pattern, and a
@@ -39,8 +41,7 @@ Added
   creation of similarity metrics used in pattern matching, (2) computation of an
   orientation similarity map from indexing results, and (3) creation of a multi
   phase crystal map from single phase maps from pattern matching.
-- EBSD.xmap property storing an orix CrystalMap object. So far only read from
-  a EMsoft simulated EBSD pattern file. Relevant documentation updated.
+- EBSD.xmap property storing an orix CrystalMap.
   (`#226 <https://github.com/pyxem/kikuchipy/pull/226>`_)
 - Dependency on the diffsims package for handling of electron scattering and
   diffraction. (`#220 <https://github.com/pyxem/kikuchipy/pull/220>`_)
@@ -66,7 +67,7 @@ Changed
 - The EBSDMasterPattern gets phase, hemisphere and projection properties.
   (`#246 <https://github.com/pyxem/kikuchipy/pull/246>`_)
 - EMsoft EBSD master pattern plugin can read a single energy pattern. Parameter
-  `energy_range`  changed to `energy`.
+  `energy_range` changed to `energy`.
   (`240 <https://github.com/pyxem/kikuchipy/pull/240>`_)
 - Migrate user guide from reST files to Jupyter Notebooks converted to HTML with
   the `nbsphinx` package.